refactor: repository 추가, 안 쓰는 안터티 삭제
