+++ src/main/java/io/cavia/trader/module/game/controller/GameSessionController.java
@@ -2,7 +2,6 @@ package io.cavia.trader.module.game.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
@@ -11,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 public class GameSessionController {
 
     @RequestMapping("")
-    public String testDetailPage() {
+    public String getGameMain() {
         return "game/main";
     }
 }
