+++ src/main/java/io/cavia/trader/module/game/repository/GameMapper.java
@@ -1,7 +1,7 @@
 package io.cavia.trader.module.game.repository;
 
 import io.cavia.trader.module.game.entity.GameParticipation;
-import io.cavia.trader.module.game.entity.Member;
+import io.cavia.trader.module.member.entity.Member;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
