+++ src/main/java/io/cavia/trader/module/game/service/GameManager.java
@@ -3,12 +3,11 @@ package io.cavia.trader.module.game.service;
 import io.cavia.trader.module.client.dto.QuotesOutput;
 import io.cavia.trader.module.client.dto.TradesOutput;
 import io.cavia.trader.module.game.dto.GameDto;
-import io.cavia.trader.module.game.entity.Member;
+import io.cavia.trader.module.member.entity.Member;
 import io.jsonwebtoken.Claims;
 import org.springframework.web.socket.WebSocketSession;
 
 import java.time.LocalDateTime;
-import java.util.Deque;
 import java.util.List;
 
 public interface GameManager {
